Closed #13 Field variables appear to work now, both static and non-static, although static variables can now only be called from a class variable and not from the class itself, I need to make sure that this will work too, however I've set this up for issue #19 I've also set 'self' to a keyword in NIL. I know that self is not a keyword in Lua, but yet Lua automatically uses that var if you define a function like this: function mytab:myfunc(), and therefore it was only the most obvious thing to do.
Methods are going to be the next step... Now I am planning to make methods that way that you don't have to use ':' in stead of '.' to call them (for tables that rule will stand up, as there's no way NIL can tell the difference there), however I am not fully sure of success here, but I gotta try I suppose.... Since '.' and ':' are a very easy way to bring OOP programs in Lua into a big mess, and I would recommend against using tables in a NIL script for OOP, you can prevent some trouble here, but NO PROMISES YET! Some study I did for this proved to be hopeful, but it's all in a state, that I cannot say how this will work. Oh, that functionality WILL require Lua 5.1, as Lua 5.0 cannot provide the functionality I need for this.